+++ packages/wallet-adapter-core/src/constants.ts
@@ -19,9 +19,24 @@ export const ChainIdToAnsSupportedNetworkMap: Record<string, string> = {
   "2": "testnet", // testnet
 };
 
-/** The base URL for all Aptos Connect wallets. */
+/**
+ * The base URL for all Aptos Connect wallets.
+ *
+ * @deprecated Use {@link PETRA_WEB_BASE_URL} instead.
+ */
 export const APTOS_CONNECT_BASE_URL = "https://aptosconnect.app";
 
-/** The URL to the Aptos Connect account page if the user is signed in to Aptos Connect. */
+/** The base URL for all Petra Web wallets. */
+export const PETRA_WEB_BASE_URL = "https://web.petra.app";
+
+/**
+ * The URL to the Aptos Connect account page if the user is signed in to Aptos Connect.
+ *
+ * @deprecated Use {@link PETRA_WEB_ACCOUNT_URL} instead.
+ */
 export const APTOS_CONNECT_ACCOUNT_URL =
   "https://aptosconnect.app/dashboard/main-account";
+
+/** The URL to the Petra Web account page if the user is signed in to Petra Web. */
+export const PETRA_WEB_ACCOUNT_URL =
+  "https://web.petra.app/dashboard/main-account";